-- Added mvc controller advice class. -- Removed repeated code.
