+++ src/main/java/guru/springframework/msscbrewery/web/controller/CustomerController.java
@@ -6,7 +6,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,10 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 /**
@@ -66,17 +62,4 @@ public class CustomerController {
     customerService.deleteById(customerId);
   }
 
-  @ExceptionHandler(ConstraintViolationException.class)
-  public ResponseEntity<List> validationErrorHandler(ConstraintViolationException e) {
-    List<String> errors = new ArrayList<>(e.getConstraintViolations().size());
-
-    e.getConstraintViolations()
-        .forEach(
-            constraintViolation -> {
-              errors.add(
-                  constraintViolation.getPropertyPath() + ": " + constraintViolation.getMessage());
-            });
-
-    return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
-  }
 }
