+++ src/main/java/guru/springframework/msscbrewery/web/controller/MvcControllerAdvice.java
@@ -1,0 +1,38 @@
+package guru.springframework.msscbrewery.web.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import javax.validation.ConstraintViolationException;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @author Luis M. Kupferberg Ruiz (lkupferberg@overactive.com)
+ * @created 2020-07-10 22:01
+ */
+@ControllerAdvice
+public class MvcControllerAdvice {
+
+  @ExceptionHandler(ConstraintViolationException.class)
+  public ResponseEntity<List> validationErrorHandler(ConstraintViolationException e) {
+    List<String> errors = new ArrayList<>(e.getConstraintViolations().size());
+
+    e.getConstraintViolations()
+        .forEach(
+            constraintViolation -> {
+              errors.add(
+                  constraintViolation.getPropertyPath() + ": " + constraintViolation.getMessage());
+            });
+
+    return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
+  }
+
+  @ExceptionHandler(BindException.class)
+  public ResponseEntity<List> handleBindException(BindException e) {
+    return new ResponseEntity<>(e.getAllErrors(), HttpStatus.BAD_REQUEST);
+  }
+}
