+++ src/main/java/guru/springframework/msscbrewery/web/controller/v2/BeerControllerV2.java
@@ -65,18 +65,4 @@ public class BeerControllerV2 {
   public void deleteById(@PathVariable("beerId") UUID beerId) {
     beerServiceV2.deleteById(beerId);
   }
-
-  @ExceptionHandler(ConstraintViolationException.class)
-  public ResponseEntity<List> validationErrorHandler(ConstraintViolationException e) {
-    List<String> errors = new ArrayList<>(e.getConstraintViolations().size());
-
-    e.getConstraintViolations()
-        .forEach(
-            constraintViolation -> {
-              errors.add(
-                  constraintViolation.getPropertyPath() + ": " + constraintViolation.getMessage());
-            });
-
-    return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
-  }
 }
